Renamed some core classes
